Xylem Conduit Scaling Data

A data frame used as an example in balancR, representing simulated xylem conduit measurements. This data set was used to generate Fig. 3 in doi:10.1111/geb.70019 .

Usage

xylem_scaling_simulation_dataset

Format

A data frame with 441,904 rows and 3 variables:

Organ

Plant organ in which xylem conduits were measured (e.g., stem, leaf)

L

Distance from leaf tip (m)

DAVG

Xylem conduit diameter (µm)

Source

Generated via data-raw/build_xylem_scaling_dataset.R.